% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{addProbabilities}
\alias{addProbabilities}
\title{Function to add the probabilities in a vector}
\usage{
addProbabilities(probs)
}
\arguments{
\item{probs}{probabilities}
}
\value{
sum of probabilities if sucess else -1
}
\description{
Function to add the probabilities in a vector
}
\examples{
addProbabilities(c(0.6,0.2))
}
