% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adl_scoring}
\alias{adl_scoring}
\title{ADL scoring table}
\format{A 33 by 3 dataframe}
\source{
I-WOTCH clinical team at Warwick Clinical Trials Unit
}
\usage{
adl_scoring
}
\description{
ADL scoring table
}
\keyword{datasets}
