% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{attribute_trans_prob}
\alias{attribute_trans_prob}
\title{Attribute parameters to probabilities of transiiton matrix}
\usage{
attribute_trans_prob(tmat, probs)
}
\arguments{
\item{tmat}{A transition matrix in the format from the \link[mstate]{mstate} package.}

\item{probs}{list of probabilities as in the order of transitions (row wise)}
}
\value{
the transtion table with the probabilites
}
\description{
Attribute parameters to probabilities of transiiton matrix
}
\examples{
tmat<-rbind(c(1, 2),c(3, 4))
colnames(tmat)<-rownames(tmat)<-c("Healthy","Dead")
tm=attribute_trans_prob(tmat,probs=c(0.2,0.5,0,0.3))
}
