% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{check.trans.prob}
\alias{check.trans.prob}
\title{Check the tranisition probabilties row sum is 1}
\usage{
check.trans.prob(trans.mat)
}
\arguments{
\item{trans.mat}{transition matrix}
}
\value{
rowsum if they add upto 1 else error
}
\description{
Check the tranisition probabilties row sum is 1
}
\examples{
tmat<-rbind(c(1, 2),c(3, 4))
colnames(tmat)<-rownames(tmat)<-c("Healthy","Dead")
tm<-transition.matrix(2,tmat,list.prob=c(0.5,0.5,0,1))
check.trans.prob(tm)
}
