% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{checksumRowprob}
\alias{checksumRowprob}
\title{Function to check the sum of probabilites in a vector}
\usage{
checksumRowprob(avector)
}
\arguments{
\item{avector}{vector of probabilities}
}
\value{
sum of probabilities if sucess else error
}
\description{
Function to check the sum of probabilites in a vector
}
\examples{
checksumRowprob(c(0.6,0.4))
}
