% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{combine.health.state}
\alias{combine.health.state}
\title{Set the attribute for the health state}
\usage{
combine.health.state(state1, ...)
}
\arguments{
\item{state1}{object 1 of class health state}

\item{...}{any additional objects}
}
\value{
modified heatlh state
}
\description{
Set the attribute for the health state
}
\examples{
a<-health.state("IT",100,0.4,FALSE)
b<-health.state("PT",100,0.4,FALSE)
combine.health.state(a,b)
}
