% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get.var.health.state}
\alias{get.var.health.state}
\title{Get the attribute for the health state}
\usage{
get.var.health.state(state, var)
}
\arguments{
\item{state}{object of class health state}

\item{var}{attribute of the health state}
}
\value{
modified heatlh state
}
\description{
Get the attribute for the health state
}
\examples{
get.var.health.state(health.state("IT",100,0.4,FALSE),"cost")
}
