% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{health.state}
\alias{health.state}
\title{Definition of health state class or health state constructor}
\usage{
health.state(name, cost, utility, absorb = FALSE)
}
\arguments{
\item{name}{name of the health state}

\item{cost}{cost of the health state}

\item{utility}{utility of the health state}

\item{absorb}{boolean indicating health state abosribing or not}
}
\value{
value of the state
}
\description{
Definition of health state class or health state constructor
}
\examples{
st=health.state("IT",100,0.4,FALSE)
}
