% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{init.trace}
\alias{init.trace}
\title{Define an all zero trace matrix}
\usage{
init.trace(health.states, cycles)
}
\arguments{
\item{health.states}{heatlh states}

\item{cycles}{no of cycles}
}
\value{
trace matrix -all zero
}
\description{
Define an all zero trace matrix
}
\examples{
a<-health.state("Healthy",1,1,FALSE)
b<-health.state("Dead",1,0.5,FALSE)
health.states=combine.health.state(a,b)
init.trace(health.states,10)
}
