% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{markov.model}
\alias{markov.model}
\title{Definition of Markov model}
\usage{
markov.model(this.strategy, cycles, initial.state)
}
\arguments{
\item{this.strategy}{strategy object}

\item{cycles}{no of cycles}

\item{initial.state}{value of states initially}
}
\value{
Markov model
}
\description{
Definition of Markov model
}
\examples{
tmat <- rbind(c(1, 2),c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm=transition.matrix(2,tmat,c(0.5,0.5,0,1))
a<-health.state("Healthy",1,1,FALSE)
b<-health.state("Dead",1,0.5,FALSE)
health.states=combine.health.state(a,b)
this.strategy=strategy(tm,health.states,"intervention")
markov.model(this.strategy,10,c(1,0))
}
