% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_help_functions.R
\name{randomNumbersProbDistbn}
\alias{randomNumbersProbDistbn}
\title{Function to return random number given a distribution and parameters}
\usage{
randomNumbersProbDistbn(prob.distbn, param)
}
\arguments{
\item{prob.distbn}{a probabiltiy distribution and vector of parameters}

\item{param}{param values to define the probability distribution}
}
\value{
random number according to the specified distribution
}
\description{
Function to return random number given a distribution and parameters
}
\examples{
randomNumbersProbDistbn("beta",c(0.6,0.2))
}
