% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{set.var.health.state}
\alias{set.var.health.state}
\title{Set the attribute for the health state}
\usage{
set.var.health.state(state, var, new.value)
}
\arguments{
\item{state}{object of class health state}

\item{var}{attribute of the health state}

\item{new.value}{new value to be assigned}
}
\value{
modified heatlh state
}
\description{
Set the attribute for the health state
}
\examples{
set.var.health.state(health.state("IT",100,0.4,FALSE),"cost",1)
}
