% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{strategy}
\alias{strategy}
\title{Definition of strategy -or arm}
\usage{
strategy(trans.mat, states, name)
}
\arguments{
\item{trans.mat}{tranisiton matrix}

\item{states}{heatlh states}

\item{name}{name of the strategy}
}
\value{
object strategy
}
\description{
Definition of strategy -or arm
}
\examples{
tmat <- rbind(c( 1, 2),c(3, 4))
colnames(tmat) <- rownames(tmat) <- c("Healthy", "Dead")
tm=transition.matrix(2,tmat,c(0.5,0.5,0,1))
a<-health.state("Healthy",1,1,FALSE)
b<-health.state("Dead",1,0.5,FALSE)
states=combine.health.state(a,b)
strategy(tm,states,"intervention")
}
