% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{transition.matrix}
\alias{transition.matrix}
\title{Create the tranistion matrix}
\usage{
transition.matrix(no.states, tmat, list.prob, name.states = NULL)
}
\arguments{
\item{no.states}{number of the health states}

\item{tmat}{A transition matrix in the format from the \link[mstate]{mstate} package.}

\item{list.prob}{list of probabilities as in the order of transitions (row wise)}

\item{name.states}{names of the health states}
}
\value{
value of the transition matrix
}
\description{
Create the tranistion matrix
}
\examples{
tmat<-rbind(c(1, 2),c(3, 4))
colnames(tmat)<-rownames(tmat)<-c("Healthy","Dead")
transition.matrix(2,tmat,list.prob=c(0.2,0.5,0,0.3))
}
