% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial.data}
\alias{trial.data}
\title{Trial data}
\format{A 34 by 17 dataframe}
\source{
I-WOTCH clinical team at Warwick Clinical Trials Unit
}
\usage{
trial.data
}
\description{
Trial data
}
\keyword{datasets}
