% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial_data}
\alias{trial_data}
\title{Trial data}
\format{A 33 by 31 dataframe}
\source{
I-WOTCH clinical team at Warwick Clinical Trials Unit
}
\usage{
trial_data
}
\description{
Trial data
}
\keyword{datasets}
